requirements.txt
backoff==2.2.1
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
future==0.18.2
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.23.5
pandas==1.5.2
pathspec==0.10.2
platformdirs==2.5.4
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.10.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
webex-bot==0.3.4
webexteamssdk==1.6.1
websockets==10.2
